Move x2sys.py to src folder and split it into x2sys_init and x2sys_cross
Move the x2sys modules originally wrapped in #546.
The single x2sys.py has been separated into x2sys_init.py
and x2sys_cross.py inside the src/ folder as per #807.
