+++ pygmt/src/__init__.py
@@ -28,3 +28,5 @@ from pygmt.src.subplot import set_panel, subplot
 from pygmt.src.surface import surface
 from pygmt.src.text import text_ as text  # "text" is an argument within "text_"
 from pygmt.src.which import which
+from pygmt.src.x2sys_init import x2sys_init
+from pygmt.src.x2sys_cross import x2sys_cross
